fixes, create NftForm.utils
