+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/NftForm.hooks.ts
@@ -1,9 +1,8 @@
-import { differenceInMilliseconds, formatDuration, intervalToDuration } from 'date-fns'
+import { differenceInMilliseconds } from 'date-fns'
 import { useCallback, useState } from 'react'
 
 import { AuctionDatePickerValue } from '@/components/_inputs/AuctionDatePicker'
 import { useBlockTimeEstimation } from '@/hooks/useBlockTimeEstimation'
-import { pluralizeNoun } from '@/utils/misc'
 import { daysToMilliseconds } from '@/utils/time'
 
 import { Listing } from './NftForm.types'
@@ -15,19 +14,6 @@ export const useNftForm = () => {
   const [currentStep, setCurrentStep] = useState(0)
 
   const { convertDurationToBlocks } = useBlockTimeEstimation()
-
-  const getTotalDaysAndHours = (startDate: AuctionDatePickerValue, endDate: AuctionDatePickerValue) => {
-    const start = (startDate?.type === 'date' && startDate.date) || new Date()
-
-    if (endDate?.type === 'date') {
-      const { days, hours } = intervalToDuration({ start: start, end: endDate.date })
-      const duration = formatDuration({ hours: hours, days: days }, { format: ['days', 'hours'] })
-      return duration ? duration : 'Less than 1 hour'
-    }
-    if (endDate?.type === 'duration') {
-      return pluralizeNoun(endDate.durationDays, 'Day')
-    }
-  }
 
   const nextStep = useCallback(() => setCurrentStep((step) => step + 1), [])
   const previousStep = useCallback(() => setCurrentStep((step) => step - 1), [])
@@ -46,7 +32,6 @@ export const useNftForm = () => {
   }
 
   return {
-    getTotalDaysAndHours,
     getNumberOfBlocks,
     state: {
       activeInputs,
